pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "apitool-cli"
version = "1.0.0"
description = "Browser Recording & Security Testing CLI Tool"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "API Tool CLI Contributors", email = ""},
]
maintainers = [
    {name = "API Tool CLI Contributors", email = ""},
]
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Security",
    "Topic :: Software Development :: Testing",
    "Topic :: System :: Monitoring",
]
keywords = [
    "browser automation",
    "security testing",
    "api testing",
    "recording",
    "playwright",
    "penetration testing",
    "web scraping",
]

dependencies = [
    "playwright>=1.40.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "pyaudio>=0.2.11",
    "openai>=1.0.0",
    "numpy>=1.24.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/apitool-cli"
Documentation = "https://github.com/yourusername/apitool-cli#readme"
Repository = "https://github.com/yourusername/apitool-cli.git"
Issues = "https://github.com/yourusername/apitool-cli/issues"

[project.scripts]
apitool = "scripts.cli:cli"

[tool.setuptools]
packages = ["scripts"]

[tool.setuptools.package-data]
scripts = ["*.py"]

[tool.black]
line-length = 120
target-version = ["py38", "py39", "py310", "py311"]

[tool.flake8]
max-line-length = 120
exclude = [".git", "__pycache__", "venv", "build", "dist"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true